feat(organizeImports): restore first-specifier sorting
This is achieved by moving in-group sorting responsibility to
import groups, instead of trying to achieve this in the import
manager, only to be overridden by the groups' `sortedImports` getter
inside their `generate()` calls, as discussed in
buehler/typescript-hero#369.

This isn't marked as closing yet, as I am facing an issue with
manual testing going fine, but automated testing VSC debug target
failing to locate the IocDecorators export, apparently, thereby
b0rking on `lazyInject` for config DI.  I need input from @buehler
on this.