feat(import-management): Add imports and imports under cursor (#383)
Re-Introduce the `add import` and `add import under cursor` feature.
This feature was removed due to the fact, that it uses the declaration index.
That index was broken and caused vscode to totally crash. I removed the "support"
for javascript files, which mainly caused the crashes. This due to the fact, that
compiled javascripts (like webpack results) were indexed and crashed the memory
of the process.